Rename StageTypesDatabaseID.ts to constants.ts
